test(core): split profile composition responsibilities
